requirements.txt
click
ruamel.yaml